UserService passes all tests

+++ server/src/main/java/service/UserService.java
@@ -3,6 +3,8 @@ package service;
 import chess.*;
 import dataaccess.*;
 import model.*;
+
+import java.util.Objects;
 import java.util.UUID;
 
 public class UserService {
@@ -18,21 +20,34 @@ public class UserService {
         authDAO = authIn;
     }
 
-    public AuthData register(UserData newUser) {
-        userDAO.get(newUser.username()); // FIXME: AlreadyTakenException [403]
+    public AuthData register(UserData newUser) throws AlreadyTakenException {
+        if(userDAO.get(newUser.username()) != null) { // TODO: AlreadyTakenException [403]
+            throw new AlreadyTakenException("already taken");
+        }
         userDAO.create(newUser);
         AuthData loginInfo = new AuthData(generateToken(), newUser.username());
         authDAO.create(loginInfo);
         return loginInfo;
     }
+
     public AuthData login(LoginRequest loginRequest) {
-        userDAO.get(loginRequest.username()); // FIXME: NotFoundException [404], UnauthorizedException [401]
+        var userRetrieved = userDAO.get(loginRequest.username());
+        if(userRetrieved == null) { // TODO: NotFoundException [404]
+            throw new NotFoundException("username not found");
+        }
+        if(!userRetrieved.password().equals(loginRequest.password())) { // TODO: UnauthorizedException [401]
+            throw new UnauthorizedException("unauthorized");
+        }
         return new AuthData(generateToken(), loginRequest.username());
     }
+
     public void logout(String authToken) {
-        authDAO.get(authToken); // FIXME: UnauthorizedException [401]
+        if(authDAO.get(authToken) == null) { // TODO: UnauthorizedException [401]
+            throw new UnauthorizedException("unauthorized");
+        }
         authDAO.delete(authToken);
     }
+
     public void clear() {
         authDAO.clear();
         userDAO.clear();
